EdrServer/requirements.txt
celery~=5.4.0
Django~=5.0.6
evtx~=0.8.4
PyYAML~=6.0.1
redis~=5.0.5